refactor: extract Location field length constants
