+++ src/StuffTracker.Domain/Constants/LocationConstants.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace StuffTracker.Domain.Constants
+{
+    public static class LocationConstants
+    {
+        public const int LocationNameMaxLength = 100;
+        public const int LocationDescriptionMaxLength = 500;
+    }
+}
